requirements.txt
packaging==25.0
pillow==12.0.0
pytesseract==0.3.13
